+++ src/hooks/useEvent.js
@@ -2,7 +2,7 @@ import { useEffect } from "react";
 
 export default function useEvent(target, event, listener, passive = true) {
     useEffect(() => {
-        if (target === undefined) return;
+        if (!target) return;
         const options = { passive };
 
         target.addEventListener(event, listener, options);